IGNITE-19928 Fix method signature related to creating a new error group and registering a new error code

+++ modules/core/src/test/java/org/apache/ignite/lang/ErrorGroupTest.java
@@ -88,7 +88,7 @@ class ErrorGroupTest {
     @SuppressWarnings({"rawtypes", "OptionalGetWithoutIsPresent"})
     @Test
     void groupIdsAreUnique() throws IllegalAccessException {
-        Map<Integer, ErrorGroup> errGroups = new HashMap<>();
+        Map<Short, ErrorGroup> errGroups = new HashMap<>();
 
         for (Class cls : ErrorGroups.class.getDeclaredClasses()) {
             var errGroupField = Arrays.stream(cls.getFields()).filter(f -> f.getName().endsWith("_ERR_GROUP")).findFirst().get();
